Create ActionProvider facade for the ActionDelegateProvider, to store the necessary params for getting the action delegates.
COSDK-480
