+++ core/src/main/java/com/adyen/checkout/core/components/internal/PaymentFacilitatorProvider.kt
@@ -25,6 +25,7 @@ internal class PaymentFacilitatorProvider {
             AdvancedPaymentFacilitatorFactory(
                 checkoutConfiguration = adyenCheckout.checkoutConfiguration,
                 checkoutCallback = adyenCheckout.checkoutCallback,
+                savedStateHandle = savedStateHandle,
             )
         } else {
             SessionsPaymentFacilitatorFactory(
